Moved most content to soilmask-transformer
The soilmask transformer has been separated into 1) generate soilmask in https://github.com/AgPipeline/transformer-soilmask/ and 2) compute area of soilmask that is plant (this repository). Updating both readmes to reflect this.

TODO:

Check that this is consistent with existing code and examples (this repo should start with a mask and return a canopy cover, or else have the soilmask as a dependency and take in an RGB image and return a mask + canopy cover). See also https://github.com/AgPipeline/transformer-soilmask/pull/11