Merging develop into master as v 3.2 (#10)
* Fixing up README and type declarations (#1)

* Minor cleanup

* Updated file docstring

* Updating README

* Fixing up return type

* Ignore .idea folders, put old README contents into new README, reconciled return types

* update method name (#5)

Same as https://github.com/terraref/extractors-stereo-rgb/pull/56; to match method name in BETYdb

* Moved most content to soilmask-transformer

The soilmask transformer has been separated into 1) generate soilmask in https://github.com/AgPipeline/transformer-soilmask/ and 2) compute area of soilmask that is plant (this repository). Updating both readmes to reflect this.

TODO:

Check that this is consistent with existing code and examples (this repo should start with a mask and return a canopy cover, or else have the soilmask as a dependency and take in an RGB image and return a mask + canopy cover). See also https://github.com/AgPipeline/transformer-soilmask/pull/11

* Update README.md

Co-Authored-By: Chris Schnaufer <schnaufer@email.arizona.edu>

* moved figures to transformer-soilmask repository

* Pushing transformer.py

* tried to update algorithm for pylint preferences; might still be some errors because indentation did not seem to be preserved and wasn't present on the repo

* Changed the masked canopy cover calculation in transformer.py

* updated canopy cover algorithm and ran pylint

* made changes

* made changes and ran pylint

* changed file to reflect pull request

* Update README.md

* Update README.md

* Fixed algorithm (#9)

Co-authored-by: David LeBauer <dlebauer@arizona.edu>
Co-authored-by: David LeBauer <dlebauer@email.arizona.edu>
Co-authored-by: Jacob van der Leeuw <jacobvanderleeuw@Jacobs-MacBook-Pro.local>
Co-authored-by: jvanderleeuw <62153719+jvanderleeuw@users.noreply.github.com>