Finishing up cleanup (#14)
* Moved most content to soilmask-transformer

The soilmask transformer has been separated into 1) generate soilmask in https://github.com/AgPipeline/transformer-soilmask/ and 2) compute area of soilmask that is plant (this repository). Updating both readmes to reflect this.

TODO:

Check that this is consistent with existing code and examples (this repo should start with a mask and return a canopy cover, or else have the soilmask as a dependency and take in an RGB image and return a mask + canopy cover). See also https://github.com/AgPipeline/transformer-soilmask/pull/11

* Update README.md

Co-Authored-By: Chris Schnaufer <schnaufer@email.arizona.edu>

* moved figures to transformer-soilmask repository

* Update README.md

* Update README.md

* Fixed algorithm (#9)

* Fixing up Dockerfile

* Updated Dockerfile to recover lost code

Co-authored-by: David LeBauer <dlebauer@arizona.edu>
Co-authored-by: David LeBauer <dlebauer@email.arizona.edu>